move more functions to utility. all parameters in frameworks are now defined for all utilities (not all might be meaningful)
