+++ nmtwizard/utility.py
@@ -6,13 +6,16 @@ import argparse
 import json
 import time
 import re
+import shutil
 import uuid
 import six
 
-from nmtwizard.utils import merge_dict
 from nmtwizard.beat_service import start_beat_service
-from nmtwizard.logger import get_logger
 from nmtwizard.storage import StorageClient
+from nmtwizard.utils import md5files
+from nmtwizard.utils import merge_dict
+from nmtwizard.logger import get_logger
+from nmtwizard import data
 
 ENVVAR_RE = re.compile(r'\${(.*?)}')
 ENVVAR_ABS_RE = re.compile(r'(\${.*?}.*)/(.*)')
@@ -73,6 +76,7 @@ class Utility(object):
             os.makedirs(self._data_dir)
         if not os.path.exists(self._tmp_dir):
             os.makedirs(self._tmp_dir)
+        self._parser = argparse.ArgumentParser()
 
     @property
     @abc.abstractmethod
@@ -101,9 +105,27 @@ class Utility(object):
         parser.add_argument('-bi', '--beat_interval', default=30, type=int,
                             help="Interval of beat requests in seconds.")
 
-        parser.add_argument('utility_args', nargs=argparse.REMAINDER)
+        parser.add_argument('-ms', '--model_storage', default=None,
+                            help='Model storage in the form <storage_id>:[<path>].')
+        parser.add_argument('-msr', '--model_storage_read', default=None,
+                            help=('Model storage to read from, in the form <storage_id>:[<path>] '
+                                  '(defaults to model_storage).'))
+        parser.add_argument('-msw', '--model_storage_write', default=None,
+                            help=('Model storage to write to, in the form <storage_id>:[<path>] '
+                                  '(defaults to model_storage).'))
+        parser.add_argument('-c', '--config', default=None,
+                            help=('Configuration as a file or a JSON string. '
+                                  'Setting "-" will read from the standard input.'))
+        parser.add_argument('-m', '--model', default=None,
+                            help='Model to load.')
+        parser.add_argument('-g', '--gpuid', default="0",
+                            help="Comma-separated list of 1-indexed GPU identifiers (0 for CPU).")
+        parser.add_argument('--no_push', default=False, action='store_true',
+                            help='Do not push model.')
 
-        args, unknown = parser.parse_known_args(args=args)
+        parser.add_argument('utility_args', nargs=argparse.REMAINDER, help=self._parser.format_usage())
+
+        args = parser.parse_args(args=args)
 
         if args.task_id is None:
             args.task_id = str(uuid.uuid4())
@@ -121,11 +143,86 @@ class Utility(object):
             tmp_dir=self._tmp_dir,
             config=load_config(args.storage_config) if args.storage_config else None)
 
+        if args.model_storage_read is None:
+            args.model_storage_read = args.model_storage
+        if args.model_storage_write is None:
+            args.model_storage_write = args.model_storage
+
+        self._model_storage_read = args.model_storage_read
+        self._model_storage_write = args.model_storage_write
+
+        # for backward compatibility - convert singleton in int
+        args.gpuid = args.gpuid.split(',')
+        args.gpuid = [int(g) for g in args.gpuid]
+        if len(args.gpuid) == 1:
+            args.gpuid = args.gpuid[0]
+
+        self._gpuid = args.gpuid
+
+        self._config = load_config(args.config) if args.config is not None else None
+        self._model = args.model
+        self._no_push = args.no_push
+
         logger.info('Starting executing utility %s=%s with args [%s]',
                     self.name, args.image, " ".join(args.utility_args))
         start_time = time.time()
 
-        self.exec_function(unknown + args.utility_args)
+        self.exec_function(args.utility_args)
 
         end_time = time.time()
         logger.info('Finished executing utility in %s seconds', str(end_time-start_time))
+
+    def _merge_multi_training_files(self, data_path, train_dir, source, target):
+        merged_dir = os.path.join(self._data_dir, 'merged')
+        if not os.path.exists(merged_dir):
+            os.mkdir(merged_dir)
+        merged_path = os.path.join(merged_dir, train_dir)
+        logger.info('Merging training data to %s/train.{%s,%s}',
+                    merged_path, source, target)
+        data.merge_files_in_directory(data_path, merged_path, source, target)
+        return merged_path
+
+
+def build_model_dir(model_dir, objects, config, check_integrity_fn):
+    """Prepares the model directory based on the model package."""
+    if os.path.exists(model_dir):
+        raise ValueError("model directory %s already exists" % model_dir)
+    else:
+        logger.info("Building model package in %s", model_dir)
+    os.mkdir(model_dir)
+    for target, source in six.iteritems(objects):
+        if os.path.isdir(source):
+            shutil.copytree(source, os.path.join(model_dir, target))
+        else:
+            shutil.copyfile(source, os.path.join(model_dir, target))
+    config_path = os.path.join(model_dir, 'config.json')
+    with open(config_path, 'w') as config_file:
+        json.dump(config, config_file)
+    objects['config.json'] = config_path
+    if "description" in config:
+        readme_path = os.path.join(model_dir, 'README.md')
+        with open(readme_path, 'w') as readme_file:
+            readme_file.write(config['description'])
+        objects['README.md'] = readme_path
+    md5 = md5files((k, v) for k, v in six.iteritems(objects) if check_integrity_fn(k))
+    with open(os.path.join(model_dir, "checksum.md5"), "w") as f:
+        f.write(md5)
+
+def check_model_dir(model_dir, check_integrity_fn):
+    """Compares model package MD5."""
+    logger.info("Checking integrity of model package %s", model_dir)
+    md5_file = os.path.join(model_dir, "checksum.md5")
+    if not os.path.exists(md5_file):
+        return True
+    md5ref = None
+    with open(md5_file, "r") as f:
+        md5ref = f.read().strip()
+    files = os.listdir(model_dir)
+    md5check = md5files([(f, os.path.join(model_dir, f)) for f in files if check_integrity_fn(f)])
+    return md5check == md5ref
+
+def fetch_model(storage, remote_model_path, model_path, check_integrity_fn):
+    """Downloads the remote model."""
+    storage.get(remote_model_path, model_path, directory=True,
+                check_integrity_fn=lambda m: check_model_dir(m, check_integrity_fn))
+    os.environ['MODEL_DIR'] = model_path
